@@ -2,6 +2,7 @@
   "name": "client-app",
   "version": "0.0.1",
   "scripts": {
+    "move-scripts": "@powershell  -NoProfile -Command ./install-contracts.ps1",
     "ng": "ng",
     "start": "ng serve",
     "build": "ng build",
@@ -12,31 +13,33 @@
   },
   "private": true,
   "dependencies": {
-    "@angular/animations": "~9.1.11",
-    "@angular/common": "~9.1.11",
-    "@angular/compiler": "~9.1.11",
-    "@angular/core": "~9.1.11",
-    "@angular/forms": "~9.1.11",
-    "@angular/platform-browser": "~9.1.11",
-    "@angular/platform-browser-dynamic": "~9.1.11",
-    "@angular/router": "~9.1.11",
+    "@angular/animations": "~11.1.0",
+    "@angular/cdk": "^11.1.0",
+    "@angular/common": "~11.1.0",
+    "@angular/compiler": "~11.1.0",
+    "@angular/core": "~11.1.0",
+    "@angular/forms": "~11.1.0",
+    "@angular/material": "^11.1.0",
+    "@angular/platform-browser": "~11.1.0",
+    "@angular/platform-browser-dynamic": "~11.1.0",
+    "@angular/router": "~11.1.0",
     "ng": "0.0.0",
     "rxjs": "~6.5.4",
     "serve": "^11.3.2",
-    "tslib": "^1.10.0",
+    "tslib": "^2.0.0",
     "web3": "^1.3.3",
     "web3-eth": "^1.3.3",
     "zone.js": "~0.10.2"
   },
   "devDependencies": {
-    "@angular-devkit/build-angular": "~0.901.9",
-    "@angular/cli": "~9.1.9",
-    "@angular/compiler-cli": "~9.1.11",
+    "@angular-devkit/build-angular": "~0.1101.1",
+    "@angular/cli": "~11.1.1",
+    "@angular/compiler-cli": "~11.1.0",
     "@types/node": "^12.11.1",
-    "codelyzer": "^5.1.2",
+    "codelyzer": "^6.0.0",
     "protractor": "~7.0.0",
     "ts-node": "~8.3.0",
     "tslint": "~6.1.0",
-    "typescript": "~3.8.3"
+    "typescript": "~4.1.3"
   }
 }
