@@ -5,18 +5,51 @@ import { HttpClientModule } from '@angular/common/http';
 import { AppRoutingModule } from './app-routing.module';
 import { AppComponent } from './app.component';
 import { ContractsService } from './services/services/contracts.service';
+import { MetaMaskService } from './services/services/metamask.service';
+import { Web3Service } from './services/services/web3.service';
+import { ConnectComponent } from './pages/connect/connect.component';
+import { HomeComponent } from './pages/home/home.component';
+import { BrowserAnimationsModule } from '@angular/platform-browser/animations';
+import { HeaderComponent } from './comp/header/header.component';
+import { MatButtonModule } from '@angular/material/button';
+import { MatDividerModule } from '@angular/material/divider';
+import { MatCardModule } from '@angular/material/card';
+import { MatDialogModule } from '@angular/material/dialog';
+import { MatFormFieldModule, MAT_FORM_FIELD_DEFAULT_OPTIONS } from '@angular/material/form-field';
+import { MatInputModule } from '@angular/material/input';
+import { FormsModule, ReactiveFormsModule } from '@angular/forms';
+
+import {MatNativeDateModule} from '@angular/material/core';
+import { NewProductDialog } from './pages/home/new-product/new-product.dialog';
+
 
 @NgModule({
   declarations: [
-    AppComponent
+    AppComponent,
+    ConnectComponent,
+    HomeComponent,
+    HeaderComponent,
+    NewProductDialog
   ],
   imports: [
     BrowserModule,
     AppRoutingModule,
-    HttpClientModule
+    HttpClientModule,
+    BrowserAnimationsModule,
+    MatButtonModule,
+    MatDividerModule,
+    MatCardModule,
+    MatDialogModule,
+    MatFormFieldModule,
+    MatInputModule,
+    ReactiveFormsModule,
+    FormsModule,
+    MatNativeDateModule
   ],
   providers: [
-    ContractsService
+    ContractsService,
+    MetaMaskService,
+    Web3Service
   ],
   bootstrap: [AppComponent]
 })
