@@ -1,5 +1,5 @@
 const fs = require('fs');
-const f = 'node_modules/@angular-devkit/build-angular/src/angular-cli-files/models/webpack-configs/browser.js';
+const f = 'node_modules/@angular-devkit/build-angular/src/webpack/configs/browser.js';
 
 console.log("PATHCING ANGULAR WEB3")
 
