@@ -1,11 +1,22 @@
 import { NgModule } from '@angular/core';
 import { Routes, RouterModule } from '@angular/router';
+import { ConnectComponent } from './pages/connect/connect.component';
+import { HomeComponent } from './pages/home/home.component';
 
 
-const routes: Routes = [];
+const routes: Routes = [
+  {
+    path: "home",
+    component: HomeComponent,
+  },
+  {
+    path: "connect",
+    component: ConnectComponent
+  }
+];
 
 @NgModule({
-  imports: [RouterModule.forRoot(routes)],
+  imports: [RouterModule.forRoot(routes, { relativeLinkResolution: 'legacy' })],
   exports: [RouterModule]
 })
 export class AppRoutingModule { }
