@@ -1,23 +1,25 @@
 import { Injectable, OnInit } from '@angular/core';
 
-import YetAnotherEthereumToken from "@assets/YetAnotherEthereumToken.json";
-import Marketplace from "@assets/Main.json";
+import YetAnotherEthereumToken from "@assets/build/YetAnotherEthereumToken.json";
+import Marketplace from "@assets/build/Main.json";
+import { Web3Service } from './web3.service';
+import contracts from "@assets/contracts.json";
+import { AbiItem, Contract } from "web3-eth/lib";
 
 @Injectable({
   providedIn: 'root'
 })
-export class ContractsService{
-
-  private contracts:string[];
-  private readonly Token = "YetAnotherEthereumToken";
-
-  constructor() { }
-
-  public get YetAnotherEthereumToken() {
-    return YetAnotherEthereumToken["abi"];
+export class ContractsService {
+  constructor(private readonly w3: Web3Service) {
   }
 
-  public get Marketplace() {
-    return Marketplace["abi"];
+  public get YetAnotherEthereumToken(): Contract{
+    return new this.w3.web3.eth.Contract
+      (YetAnotherEthereumToken["abi"] as AbiItem[], contracts.YetAnotherEthereumToken);
+  }
+
+  public get Marketplace(): Contract{
+    return new this.w3.web3.eth.Contract
+      (Marketplace["abi"] as AbiItem[], contracts.Main);
   }
 }
